refactoring to new implementation

+++ internal/core/service/project/service.go
@@ -11,13 +11,13 @@ type Storage struct {
 }
 
 type Service interface {
-	RequestUploadUrl(ctx context.Context, mimeType string, fileSize int) (url string, expiredAt int, err error)
+	RequestUploadUrl(ctx context.Context, mimeType string, fileSize int64) (url string, expiredAt int64, err error)
 	Submit(ctx context.Context, rb req.SubmitProjectReqBody) (err error)
 	ReviewByAdmin(ctx context.Context, rb req.ReviewProjectReqBody) (err error)
 	Get(ctx context.Context) (err error)
-	GetById(ctx context.Context, id uint32) (err error)
-	DonateToProject(ctx context.Context, id uint32, rb req.DonateToProjectReqBody) (err error)
-	GetDonationById(ctx context.Context, id uint32) (err error)
+	GetById(ctx context.Context, id int64) (err error)
+	DonateToProject(ctx context.Context, id int64, rb req.DonateToProjectReqBody) (err error)
+	GetDonationById(ctx context.Context, id int64) (err error)
 }
 
 func NewService(storage DataStorage, fileStorage FileStorage) Service {
@@ -48,17 +48,17 @@ func (s Storage) Get(ctx context.Context) (err error) {
 	panic("implement me")
 }
 
-func (s Storage) GetById(ctx context.Context, id uint32) (err error) {
+func (s Storage) GetById(ctx context.Context, id int64) (err error) {
 	//TODO implement me
 	panic("implement me")
 }
 
-func (s Storage) DonateToProject(ctx context.Context, id uint32, rb req.DonateToProjectReqBody) (err error) {
+func (s Storage) DonateToProject(ctx context.Context, id int64, rb req.DonateToProjectReqBody) (err error) {
 	//TODO implement me
 	panic("implement me")
 }
 
-func (s Storage) GetDonationById(ctx context.Context, id uint32) (err error) {
+func (s Storage) GetDonationById(ctx context.Context, id int64) (err error) {
 	//TODO implement me
 	panic("implement me")
 }
